Add support for several decoding methods

See merge request lab/byseqlib!4

+++ model/encoder.h
@@ -1,16 +1,16 @@
 #pragma once
 
-#include <algorithm>
-#include <chrono>
-#include <cmath>
-#include <iostream>
-#include <string>
-
 #include <cublas_v2.h>
 #include <cuda.h>
 #include <cuda_runtime.h>
 #include <thrust/functional.h>
 #include <thrust/sequence.h>
+
+#include <algorithm>
+#include <chrono>
+#include <cmath>
+#include <iostream>
+#include <string>
 
 #include "src/custom/byseqlib/proto/transformer_weight.h"
 #include "src/custom/byseqlib/tools/util.h"
@@ -79,7 +79,7 @@ class Encoder {
   Encoder(int max_batch_size, const int *p_d_token_id, int *p_d_padding_mask,
           _DataType *p_d_output, const TransformerWeight<OpType_> &tw,
           cudaStream_t stream, cublasHandle_t hd);
-  int compute_buffer_bytesize();
+  long compute_buffer_bytesize();
   void init_buffer(void *pbuf);
   std::string check();
   void run_one_infer(int batch_size, int batch_seq_len);
